Pass the HttpClient and HashService directly to createPaymentsApiClient()
