+++ src/PayU/PaymentsApi/Exceptions/AuthorizationPaymentsApiClientFactoryException.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace PayU\PaymentsApi\Exceptions;
+
+use Exception;
+
+final class AuthorizationPaymentsApiClientFactoryException extends Exception
+{
+}
