--- src/PayU/PaymentsApi/Exceptions/AuthorizationPaymentsApiFactoryException.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace PayU\PaymentsApi\Exceptions;
-
-use Exception;
-
-final class AuthorizationPaymentsApiFactoryException extends Exception
-{
-}
